// helpers/schema.tsx
/**
 * This file was generated by kysely-codegen.
 * Please do not edit it manually.
 */

import type { ColumnType } from "kysely";

export type EggType = "extra_large_white" | "large_brown" | "large_white" | "large_white_flat" | "medium_white";

export type Generated<T> = T extends ColumnType<infer S, infer I, infer U>
  ? ColumnType<S, I | undefined, U>
  : ColumnType<T, T | undefined, T>;

export type Timestamp = ColumnType<Date, Date | string, Date | string>;

export interface Adjustments {
  createdAt: Generated<Timestamp>;
  damagedDelta: Generated<number>;
  eggType: EggType;
  id: string;
  isLeftover: Generated<boolean>;
  repackedDelta: Generated<number>;
  sessionId: string;
}

export interface Entries {
  createdAt: Generated<Timestamp>;
  eggsDamaged: number;
  eggsRepacked: number;
  eggType: EggType;
  id: string;
  missingEggs: Generated<number>;
  sessionId: string;
}

export interface Sessions {
  completedAt: Timestamp | null;
  createdAt: Generated<Timestamp>;
  id: string;
  isComplete: Generated<boolean>;
  leftoverCombinedLargeRepacked: Generated<number>;
}

export interface DB {
  adjustments: Adjustments;
  entries: Entries;
  sessions: Sessions;
}


// Use these for zod schema enums or other purposes, since the kysely enums are just types
export const EggTypeArrayValues: [EggType, ...EggType[]] = ["extra_large_white","large_brown","large_white","large_white_flat","medium_white"];
// Note: kysely maps the table/column/enum names from snake_case to camelCase and PascalCase. When running SQL statements, make sure to use snake_case, but TypeScript code should use camelCase or PascalCase.